Refactor realtime server helpers and add tests
